+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Fix packaging [#2](https://github.com/opendatateam/udata-ckan/pull/2)
+- Make use of [udata pytest plugin](opendatateam/udata#1400) [#3](https://github.com/opendatateam/udata-ckan/pull/3)
 
 ## 1.0.0 (2017-10-20)